fix lots of double datatype problems

+++ src/libprimer3.c
@@ -5858,198 +5858,198 @@ p3_set_gs_primer_lowercase_masking(p3_global_settings * p , int val) {
 }
 
 void
-p3_set_gs_primer_wt_tm_gt(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_tm_gt(p3_global_settings * p , double val) {
   p->p_args.weights.temp_gt = val ;
 }
 
 void
-p3_set_gs_primer_wt_tm_lt(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_tm_lt(p3_global_settings * p , double val) {
   p->p_args.weights.temp_lt = val ;
 }
 
 void
-p3_set_gs_primer_wt_gc_percent_gt(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_gc_percent_gt(p3_global_settings * p , double val) {
   p->p_args.weights.gc_content_gt = val ;
 }
 
 void
-p3_set_gs_primer_wt_gc_percent_lt(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_gc_percent_lt(p3_global_settings * p , double val) {
   p->p_args.weights.gc_content_lt = val ;
 }
 
 void
-p3_set_gs_primer_wt_size_lt(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_size_lt(p3_global_settings * p , double val) {
   p->p_args.weights.length_lt = val ;
 }
 
 void
-p3_set_gs_primer_wt_size_gt(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_size_gt(p3_global_settings * p , double val) {
   p->p_args.weights.length_gt = val ;
 }
 
 void
-p3_set_gs_primer_wt_compl_any(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_compl_any(p3_global_settings * p , double val) {
   p->p_args.weights.compl_any = val ;
 }
 
 void
-p3_set_gs_primer_wt_compl_end(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_compl_end(p3_global_settings * p , double val) {
   p->p_args.weights.compl_end = val ;
 }
 
 void
-p3_set_gs_primer_wt_num_ns(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_num_ns(p3_global_settings * p , double val) {
   p->p_args.weights.num_ns = val ;
 }
 
 void
-p3_set_gs_primer_wt_rep_sim(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_rep_sim(p3_global_settings * p , double val) {
   p->p_args.weights.repeat_sim = val ;
 }
 
 void
-p3_set_gs_primer_wt_seq_qual(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_seq_qual(p3_global_settings * p , double val) {
   p->p_args.weights.seq_quality = val ;
 }
 
 void
-p3_set_gs_primer_wt_end_qual(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_end_qual(p3_global_settings * p , double val) {
   p->p_args.weights.end_quality = val ;
 }
 
 void
-p3_set_gs_primer_wt_pos_penalty(p3_global_settings * p , int val){
+p3_set_gs_primer_wt_pos_penalty(p3_global_settings * p , double val){
   p->p_args.weights.pos_penalty = val ;
 }
 
 void
-p3_set_gs_primer_wt_end_stability(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_end_stability(p3_global_settings * p , double val) {
   p->p_args.weights.end_stability = val ;
 }
 
 void
-p3_set_gs_primer_wt_template_mispriming(p3_global_settings * p , int val) {
+p3_set_gs_primer_wt_template_mispriming(p3_global_settings * p , double val) {
   p->p_args.weights.template_mispriming = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_tm_gt(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_tm_gt(p3_global_settings * p , double val) {
   p->o_args.weights.temp_gt = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_tm_lt(p3_global_settings * p , int val){
+p3_set_gs_primer_io_wt_tm_lt(p3_global_settings * p , double val){
   p->o_args.weights.temp_lt = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_gc_percent_gt(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_gc_percent_gt(p3_global_settings * p , double val) {
   p->o_args.weights.gc_content_gt = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_gc_percent_lt(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_gc_percent_lt(p3_global_settings * p , double val) {
   p->o_args.weights.gc_content_lt = val ;
 
 }
 
 void
-p3_set_gs_primer_io_wt_size_lt(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_size_lt(p3_global_settings * p , double val) {
   p->o_args.weights.length_lt = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_size_gt(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_size_gt(p3_global_settings * p , double val) {
   p->o_args.weights.length_gt = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_wt_compl_any(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_wt_compl_any(p3_global_settings * p , double val) {
   p->o_args.weights.compl_any = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_compl_end(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_compl_end(p3_global_settings * p , double val) {
   p->o_args.weights.compl_end = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_num_ns(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_num_ns(p3_global_settings * p , double val) {
   p->o_args.weights.num_ns = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_rep_sim(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_rep_sim(p3_global_settings * p , double val) {
   p->o_args.weights.repeat_sim = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_seq_qual(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_seq_qual(p3_global_settings * p , double val) {
   p->o_args.weights.seq_quality = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_end_qual(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_end_qual(p3_global_settings * p , double val) {
   p->o_args.weights.end_quality = val ;
 }
 
 void
-p3_set_gs_primer_io_wt_template_mishyb(p3_global_settings * p , int val) {
+p3_set_gs_primer_io_wt_template_mishyb(p3_global_settings * p, double val) {
   p->o_args.weights.template_mispriming = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_pr_penalty(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_pr_penalty(p3_global_settings * p , double val) {
   p->pr_pair_weights.primer_quality = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_io_penalty(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_io_penalty(p3_global_settings * p , double val) {
   p->pr_pair_weights.io_quality = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_diff_tm(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_diff_tm(p3_global_settings * p , double val) {
   p->pr_pair_weights.diff_tm = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_compl_any(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_compl_any(p3_global_settings * p , double val) {
   p->pr_pair_weights.compl_any = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_compl_end(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_compl_end(p3_global_settings * p , double val) {
   p->pr_pair_weights.compl_end = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_product_tm_lt(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_product_tm_lt(p3_global_settings * p , double val) {
   p->pr_pair_weights.product_tm_lt = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_product_tm_gt(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_product_tm_gt(p3_global_settings * p , double val) {
   p->pr_pair_weights.product_tm_gt = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_product_size_gt(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_product_size_gt(p3_global_settings * p , double val) {
   p->pr_pair_weights.product_size_gt = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_product_size_lt(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_product_size_lt(p3_global_settings * p , double val) {
   p->pr_pair_weights.product_size_lt = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_rep_sim(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_rep_sim(p3_global_settings * p , double val) {
   p->pr_pair_weights.repeat_sim = val ;
 }
 
 void
-p3_set_gs_primer_pair_wt_template_mispriming(p3_global_settings * p , int val) {
+p3_set_gs_primer_pair_wt_template_mispriming(p3_global_settings * p , double val) {
   p->pr_pair_weights.template_mispriming = val ;
 }
